pyproject.toml
[project]
name = "agent-lab"
version = "0.1.0"
description = "arri"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.1,<0.4.0)",
    "langchain-openai (>=0.2.0,<0.3.0)",
    "crewai (>=0.121.1,<0.122.0)",
    "langgraph (>=0.4.7,<0.5.0)",
    "openai (>=1.68.2,<1.76.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
